Implementadas todas las clases, falta ataqueEspecial y transformaciones especiales

+++ src/dragonalgoball/Modo.java
@@ -3,7 +3,7 @@ package dragonalgoball;
  
 public  abstract class Modo {
 	
-	protected double poderDePelea;
+	protected int poderDePelea;
 	protected int distanciaDeAtaque;
 	protected int velocidad;
 	
@@ -11,7 +11,7 @@ public  abstract class Modo {
 		 
 	}
 	
-	public double obtenerPoderDePelea(){
+	public int obtenerPoderDePelea(){
 		return this.poderDePelea;
 	}
 	public int obtenerDistanciaDeAtaque(){
